moved TickCounter to separate file as test-helper
